session stress testing
git-svn-id: svn://svn.caucho.com/resin/trunk@6090 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/admin/AverageSample.java
@@ -35,5 +35,5 @@ public interface AverageSample {
   /**
    * Adds a data sample to the probe
    */
-  public void add(double value);
+  public void add(long value);
 }
